Refactor StudioPage Studio Information dialog
